requirements.txt
pytorch-ignite
rich
torch
typing-extensions
